requirements.txt
alembic==1.5.4
aniso8601==8.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dataclasses==0.6
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.20.0
pandas==1.1.5
pandas-datareader==0.9.0
Pillow==8.1.0
plotly==4.14.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.23
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
